requirements.txt
appnope==0.1.0
avro==1.8.1
backports-abc==0.4
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.4.0.2
boto==2.42.0
cached-property==1.3.0
click==6.6
confluent-kafka==0.9.1.2
confluent-schemaregistry==0.1.1
cycler==0.10.0
decorator==4.0.10
extras==1.0.0
fastavro==0.10.2
fixtures==3.0.0
Flask==0.11.1
funcsigs==1.0.2
fuzzywuzzy==0.11.1
importlib==1.0.3
ipython==5.1.0
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
kazoo==2.2.1
lettuce==0.2.23
linecache2==1.0.0
lockfile==0.12.2
loremipsum==1.0.5
luigi==1.2.1
MarkupSafe==0.23
matplotlib==1.5.1
mock==2.0.0
mpld3==0.2
nose==1.3.7
numpy==1.11.1
pandas==0.18.1
pathlib2==2.1.0
pbr==1.10.0
pexpect==4.2.0
pickleshare==0.7.4
prettyconf==1.2.3
prompt-toolkit==1.0.5
psycopg2==2.6.2
ptyprocess==0.5.1
Pygments==2.1.3
pyparsing==2.1.4
python-daemon==2.1.1
python-dateutil==2.5.3
python-mimeparse==1.5.2
python-subunit==1.2.0
pytz==2016.6.1
PyYAML==3.11
requests==2.10.0
scikit-learn==0.17.1
scipy==0.18.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
sure==1.4.0
testtools==2.2.0
tornado==4.4.1
traceback2==1.4.0
traitlets==4.2.2
unittest2==1.1.0
wcwidth==0.1.7
Werkzeug==0.11.10
zc.thread==1.0.0
zc.zk==2.1.0
